+++ common/karma/data.ts
