+++ common/karma/my-property.ts
