+++ common/karma/my-staking.ts
