+++ common/karma/property-balance.ts
