+++ common/karma/property-lockup.ts
