+++ common/karma/property.ts
