+++ common/utils.ts
@@ -1,6 +1,7 @@
 import { AxiosResponse } from 'axios'
 import axios from 'axios'
 import { PostError } from './error'
+import urljoin from 'url-join'
 
 export function isNotEmpty(str: string): boolean {
 	if (typeof str === 'undefined') {
@@ -55,4 +56,21 @@ export function hasuraDataHeader(): Record<string, unknown> {
 		'x-hasura-role': process.env.HASERA_ROLE,
 		'x-hasura-admin-secret': process.env.HASURA_SECRET!,
 	}
+}
+
+function hasuraUrl(version: string): string {
+	return urljoin(process.env.HASERA_REQUEST_DESTINATION!, version, 'graphql')
+}
+
+export async function postHasura(version: string, query: string): Promise<any> {
+	const body = {
+		query: query,
+	}
+	const res = await post(hasuraUrl(version), body, hasuraDataHeader())
+	const data = JSON.parse(res.data)
+	if (data.errors) {
+		throw new Error(data.errors[0].message)
+	}
+
+	return data.data
 }
